video 504

+++ spring-crm-rest-security/spring-crm-rest-security/src/main/java/com/luv2code/springdemo/rest/CustomerRestController.java
@@ -1,7 +1,6 @@
 package com.luv2code.springdemo.rest;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -18,42 +17,94 @@ import com.luv2code.springdemo.service.CustomerService;
 
 @RestController
 @RequestMapping("/api")
-public class CostumerRestController {
+public class CustomerRestController {
 
+	// autowire the CustomerService
 	@Autowired
 	private CustomerService customerService;
-
+	
+	// add mapping for GET /customers
 	@GetMapping("/customers")
 	public List<Customer> getCustomers() {
+		
 		return customerService.getCustomers();
+		
 	}
-
+	
+	// add mapping for GET /customers/{customerId}
+	
 	@GetMapping("/customers/{customerId}")
 	public Customer getCustomer(@PathVariable int customerId) {
+		
 		Customer theCustomer = customerService.getCustomer(customerId);
-		return Optional.ofNullable(theCustomer)
-				.orElseThrow(() -> new CustomerNotFoundException("Customer id not found - " + customerId));
+		
+		if (theCustomer == null) {
+			throw new CustomerNotFoundException("Customer id not found - " + customerId);
+		}
+		
+		return theCustomer;
 	}
-
+	
+	// add mapping for POST /customers  - add new customer
+	
 	@PostMapping("/customers")
 	public Customer addCustomer(@RequestBody Customer theCustomer) {
+		
+		// also just in case the pass an id in JSON ... set id to 0
+		// this is force a save of new item ... instead of update
+		
 		theCustomer.setId(0);
+		
 		customerService.saveCustomer(theCustomer);
+		
 		return theCustomer;
 	}
-
+	
+	// add mapping for PUT /customers - update existing customer
+	
 	@PutMapping("/customers")
 	public Customer updateCustomer(@RequestBody Customer theCustomer) {
+		
 		customerService.saveCustomer(theCustomer);
+		
 		return theCustomer;
+		
 	}
-
+	
+	// add mapping for DELETE /customers/{customerId} - delete customer
+	
 	@DeleteMapping("/customers/{customerId}")
 	public String deleteCustomer(@PathVariable int customerId) {
-		Optional.ofNullable(customerService.getCustomer(customerId))
-				.orElseThrow(() -> new CustomerNotFoundException("Customer id not found - " + customerId));
 		
+		Customer tempCustomer = customerService.getCustomer(customerId);
+		
+		// throw exception if null
+		
+		if (tempCustomer == null) {
+			throw new CustomerNotFoundException("Customer id not found - " + customerId);
+		}
+				
 		customerService.deleteCustomer(customerId);
-		return "Deleted custorme id = " + customerId;
+		
+		return "Deleted customer id - " + customerId;
 	}
+	
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
